move handleClose and handleMenuOpen callbacks to useGridAutocompleteState
